feat(diagnostics): move raw-frame capture to Test Centre, add Clear
Raw-frame capture (toggle, export, matched-pair export) moves out of
Settings into the Test Centre diagnostic tools card on both platforms —
it was the last diagnostic control still living in Settings instead of
the single home the Test Centre is meant to be. Export raw sensor data
(CSV) and Recalibrate stay in Settings unchanged.

Adds a "Clear captured frames" button (with confirmation) so a reporter
can wipe an unrelated capture, reproduce a bug, then export a clean
sample. iOS lazily reopens a fresh session file on the next captured
frame; Android reopens immediately if a session was active, since its
capture file is a persistent append-only log rather than one-per-launch.
