+++ android/app/src/main/res/values/strings.xml
@@ -852,11 +852,15 @@
     <string name="l10n_test_centre_screen_android_runs_this_via_workmanager_doze_875e499e">Android runs this via WorkManager (Doze may delay it).</string>
     <string name="l10n_test_centre_screen_calibrating_validcount_hrvreadiness_min_nights_nights_c5b21706">Calibrating (%1$s/%2$s nights)</string>
     <string name="l10n_test_centre_screen_cancel_77dfd213">Cancel</string>
+    <string name="l10n_test_centre_screen_clear_719ea396">Clear</string>
+    <string name="l10n_test_centre_screen_clear_captured_frames_51b28d22">Clear captured frames</string>
+    <string name="l10n_test_centre_screen_clear_captured_frames_a217550e">Clear captured frames?</string>
     <string name="l10n_test_centre_screen_daily_auto_export_02ed9f75">Daily auto-export</string>
     <string name="l10n_test_centre_screen_debug_logging_daaa7d74">Debug logging</string>
     <string name="l10n_test_centre_screen_diagnostic_tools_04ba4d3f">Diagnostic tools</string>
     <string name="l10n_test_centre_screen_experimental_algorithms_e09581e2">Experimental algorithms</string>
     <string name="l10n_test_centre_screen_export_f3e4fadb">Export</string>
+    <string name="l10n_test_centre_screen_export_raw_log_matched_pair_d65390bf">Export raw + log (matched pair)</string>
     <string name="l10n_test_centre_screen_export_time_2ca90c2e">Export time</string>
     <string name="l10n_test_centre_screen_heads_up_this_test_mode_is_8b82ed69">Heads up: this test mode is off, so the report has no capture for it. For a </string>
     <string name="l10n_test_centre_screen_hr_from_ppg_sub_lag_interpolation_a3ed1536">HR-from-PPG sub-lag interpolation (v26 gap-fill)</string>
@@ -866,14 +870,19 @@
     <string name="l10n_test_centre_screen_recalibrate_aaa989ea">Recalibrate</string>
     <string name="l10n_test_centre_screen_recalibrate_charge_baseline_52a05a26">Recalibrate Charge baseline</string>
     <string name="l10n_test_centre_screen_recalibrate_your_charge_baseline_018e3846">Recalibrate your Charge baseline?</string>
+    <string name="l10n_test_centre_screen_record_5_mg_raw_capture_9354fe89">Record raw frames to a file</string>
+    <string name="l10n_test_centre_screen_record_5_mg_raw_capture_research_1d966bbf">Record raw frames to a file</string>
+    <string name="l10n_test_centre_screen_records_the_raw_frames_of_each_98a284df">Saves every raw frame your strap sends (WHOOP 4.0 or 5/MG, with a timestamp and the live heart rate) to a file you can share for diagnostics. This only records frames the strap already sent (it never writes to your strap), so it is safe to leave on during a history sync. Export the file and attach it to a bug report.</string>
     <string name="l10n_test_centre_screen_report_a_bug_with_my_log_5101ee49">Report a bug with my log</string>
     <string name="l10n_test_centre_screen_report_ee45c303">Report</string>
     <string name="l10n_test_centre_screen_review_before_sharing_d7050383">Review before sharing</string>
     <string name="l10n_test_centre_screen_share_09ca55ca">Share</string>
+    <string name="l10n_test_centre_screen_share_5_mg_capture_for_the_e41ac6bd">Share raw capture</string>
     <string name="l10n_test_centre_screen_share_strap_log_for_bug_reports_b9802500">Share strap log (for bug reports)</string>
     <string name="l10n_test_centre_screen_test_centre_37b36828">Test Centre</string>
     <string name="l10n_test_centre_screen_test_modes_e21f1d3c">Test modes</string>
     <string name="l10n_test_centre_screen_this_is_exactly_what_your_report_77278bdd">This is exactly what your report will contain. Nothing leaves this phone until you tap Share.</string>
+    <string name="l10n_test_centre_screen_this_deletes_every_frame_captured_0ed8f5ee">This deletes every frame captured so far. Capture keeps running if the toggle is still on.</string>
     <string name="l10n_test_centre_screen_this_restarts_the_roughly_4_night_33cce377">This restarts the roughly 4-night build-up for Charge and your HRV baseline. Your history stays.</string>
     <string name="l10n_time_of_day_background_atmospherephase_5ce21300">atmospherePhase</string>
     <string name="l10n_today_screen_arrange_today_6b699147">Arrange Today</string>
@@ -1423,7 +1432,6 @@
     <string name="l10n_settings_screen_experimental_whoop_5_mg_41ef7041">Experimental · WHOOP 5 / MG</string>
     <string name="l10n_settings_screen_export_0a116345">Export…</string>
     <string name="l10n_settings_screen_export_csv_6bce63a3">Export CSV…</string>
-    <string name="l10n_settings_screen_export_raw_log_matched_pair_d65390bf">Export raw + log (matched pair)</string>
     <string name="l10n_settings_screen_export_raw_sensor_data_csv_a171b81e">Export raw sensor data (CSV)</string>
     <string name="l10n_settings_screen_extends_the_sky_behind_the_whole_39bb82cc">Extends the sky behind the whole Today screen, so lowering Card transparency lets it show through every card. Needs the day-cycle background on.</string>
     <string name="l10n_settings_screen_feel_the_current_time_as_a_8ca41db1">Feel the current time as a sequence of buzzes (#460). Does nothing unless your strap is connected.</string>
@@ -1467,9 +1475,6 @@
     <string name="l10n_settings_screen_recalibrate_charge_baseline_52a05a26">Recalibrate Charge baseline</string>
     <string name="l10n_settings_screen_recalibrate_your_charge_baseline_018e3846">Recalibrate your Charge baseline?</string>
     <string name="l10n_settings_screen_recent_changes_and_what_to_expect_3ceb660b">Recent changes and what to expect</string>
-    <string name="l10n_settings_screen_record_5_mg_raw_capture_9354fe89">Record raw frames to a file</string>
-    <string name="l10n_settings_screen_record_5_mg_raw_capture_research_1d966bbf">Record raw frames to a file</string>
-    <string name="l10n_settings_screen_records_the_raw_frames_of_each_98a284df">Saves every raw frame your strap sends (WHOOP 4.0 or 5/MG, with a timestamp and the live heart rate) to a file you can share for diagnostics. This only records frames the strap already sent (it never writes to your strap), so it is safe to leave on during a history sync. Export the file and attach it to a bug report.</string>
     <string name="l10n_settings_screen_remove_photo_c8f5eda8">Remove photo</string>
     <string name="l10n_settings_screen_rename_d3f4cb89">Rename</string>
     <string name="l10n_settings_screen_rename_your_strap_s_bluetooth_name_6032668b">Rename your strap\'s Bluetooth name, useful for a second-hand band. The strap </string>
@@ -1482,7 +1487,6 @@
     <string name="l10n_settings_screen_set_up_automatic_backups_00b4780c">Set up automatic backups</string>
     <string name="l10n_settings_screen_settings_c7f73bb5">Settings</string>
     <string name="l10n_settings_screen_sex_e301dd60">Sex</string>
-    <string name="l10n_settings_screen_share_5_mg_capture_for_the_e41ac6bd">Share raw capture</string>
     <string name="l10n_settings_screen_share_on_device_signals_with_the_b3fd747e">Share on-device signals with the Coach</string>
     <string name="l10n_settings_screen_share_strap_log_for_bug_reports_b9802500">Share strap log (for bug reports)</string>
     <string name="l10n_settings_screen_shows_a_soft_sunrise_day_dusk_2d20b417">Shows a soft sunrise, day, dusk and night scene behind the Today screen. Turn it off for a plain dark canvas. Your cards stay exactly as readable.</string>
